Overloaded Map in DbOptionMapper.cs insead of using DbQuestionOptionMapper.cs
