+++ src/SurveyService.Mappers/Db/Interfaces/IDbOptionMapper.cs
@@ -1,6 +1,7 @@
 ﻿using LT.DigitalOffice.Kernel.Attributes;
 using LT.DigitalOffice.SurveyService.Models.Db;
 using LT.DigitalOffice.SurveyService.Models.Dto.Requests.Option;
+using System;
 
 namespace LT.DigitalOffice.SurveyService.Mappers.Db.Interfaces;
 
@@ -8,4 +9,5 @@ namespace LT.DigitalOffice.SurveyService.Mappers.Db.Interfaces;
 public interface IDbOptionMapper
 {
   DbOption Map(CreateOptionRequest request);
+  DbOption Map(CreateQuestionOptionRequest request, Guid questionId);
 }
