--- src/SurveyService.Mappers/Db/Interfaces/IDbQuestionOptionMapper.cs
@@ -1,12 +1,0 @@
-﻿using LT.DigitalOffice.Kernel.Attributes;
-using LT.DigitalOffice.SurveyService.Models.Db;
-using LT.DigitalOffice.SurveyService.Models.Dto.Requests.Option;
-using System;
-
-namespace LT.DigitalOffice.SurveyService.Mappers.Db.Interfaces;
-
-[AutoInject]
-public interface IDbQuestionOptionMapper
-{
-  DbOption Map(CreateQuestionOptionRequest request, Guid questionId);
-}
